==> 2.WebMD_Cleaning_Data.ipynb <==
<b>

<p>
<center>
<font size="6">
Data Mining Project 3
</font>
</center>
</p>

<p>
<center>
<font size="5">
WebMD Cleaning & Preparing Process
</font>
</center>
</p>

<p>
<center>
<font size="3">
Data Mining, Columbian College of Arts & Sciences, George Washington University
</font>
</center>
</p>

<p>
<center>
<font size="3">
Author: Voratham Tiabrat
</font>
</center>
</p>

</b>

# Part 2 : WebMD Cleaning & Preparing Process 

## Cleaning Process 

import pandas as pd
import numpy as np
from nltk import tokenize

Drug_clean = pd.read_csv('Drug.csv')
Drug_clean.head(10)

# Remove 'Reviews' word from the Reviews column
for i in range (Drug_clean.shape[0]):
    Drug_clean.loc[i,'Reviews'] = Drug_clean.loc[i,'Reviews'][0:len(Drug_clean.loc[i,'Reviews'])-8]

## Preparing Process 

### 1. Infomation 

### 1.1 Price 

# Get Price from information
Price = []
for i in range (Drug_clean.shape[0]):
    para = Drug_clean.iloc[i,8]
    sent = tokenize.sent_tokenize(para)
    for x in sent :
        if 'average' in x :
            words = x.split()
    for x in words :
        if '$' in x:
            z = x.replace("$", "")
            z = z[0:len(z)-1]
            Price.append(z)

### 1.2 Form of drug 

# Try to find all form of drug
count = []
all_form = ['tablets','Capsule(s)','Capsule','Tablet(s)','Tablet','tablets','Bottle','Vial(s)','Vial','Reconstituted(s)','Reconstituted','Tube','Jar','Can','Box','Syringe']
for i in range (Drug_clean.shape[0]):
    para = Drug_clean.iloc[i,8]
    sent = tokenize.sent_tokenize(para)
    for x in sent :
        if 'average' in x :
            words = x.split()
    for word in words:
        word = word.replace(",", "")
        # count collect all index that can collect form of drug as we know in all_form variables
        if word in all_form :
            count.append(i)
            break

# Find number of index that didn't collect the form of drug and find the new form
for i in range (Drug_clean.shape[0]):
    if i not in count:
        print(i)

#Use index from check and check the sentence that didn't collect the form of frug
i = 667
para = Drug_clean.iloc[i,8]
sent = tokenize.sent_tokenize(para)
for x in sent :
    if 'average' in x :
        words = x.split()
        for word in words:
            word = word.replace(",", "")
            print(word)

Drug_clean.iloc[667]

# Get Form of Drug from information

#Get all type of form by above process
all_form = ['Capsule(s)','Capsule','Tablet(s)','Tablet','tablets','Bottle','Vial(s)','Vial','Reconstituted(s)','Reconstituted','Tube','Jar','Can','Box','Syringe','Implant','Package','Pen(s)','Inhaler']
Form = []
for i in range (Drug_clean.shape[0]):
    para = Drug_clean.iloc[i,8]
    sent = tokenize.sent_tokenize(para)
    for x in sent :
        if 'average' in x :
            words = x.split()
    for word in words:
        word = word.replace(",", "")
        if word in all_form :
            Form.append(word)
            break

len(Form)

### count frequency of each form
from collections import Counter
counts = Counter(Form)
print(counts)
len(counts)

# 
tablet = ['Tablet(s)','Tablet','tablets']
capsule = ['Capsule(s)','Capsule']
cream = ['Tube','Can','Jar']
liquid_drink = ['Bottle']
liquid_inject = ['Vial','Reconstituted','Reconstituted(s)','Vial(s)','Pen(s)','Syringe']
other = ['Box','Package','Implant','Inhaler']
for i in range (len(Form)):
    if Form[i] in tablet:
        Form[i] = 'Tablet'
    elif Form[i] in capsule:
        Form[i] = 'Capsule'
    elif Form[i] in cream:
        Form[i] = 'Cream'
    elif Form[i] in liquid_drink :
        Form[i] = 'Liquid (Drink)'
    elif Form[i] in liquid_inject :
        Form[i] = 'Liquid (Inject)'
    elif Form[i] in other:
        Form[i] = 'Other'

Form[0:10]

## 2. Concatenating 

# Deleting Informtion columns
Drug_clean = Drug_clean.drop(['Information'], axis=1)
Drug_clean.head(10)

join = {'Form':Form,'Price':Price}
df2 = pd.DataFrame(join)
df2.shape

Data = pd.concat([Drug_clean,df2], axis=1)
Data.head(10)

#Change String to Float
Data['Reviews'] = Data['Reviews'].astype(float)
Data['Effective'] = Data['Effective'].astype(float)
Data['EaseOfUse'] = Data['EaseOfUse'].astype(float)
Data['Satisfaction'] = Data['Satisfaction'].astype(float)
Data['Price'] = Data['Price'].astype(float)

#Delete duplicates data
df = pd.pivot_table(Data, index=['Condition','Drug','Indication','Type','Form'],aggfunc='mean')
Data2 = pd.DataFrame()
for i in range (len(df.index)):
        temp = {'Condition': df.index[i][0],'Drug':df.index[i][1],'Indication':df.index[i][2],'Type':df.index[i][3],'Form':df.index[i][4],'EaseOfUse':df.iloc[i,0],'Effective':df.iloc[i,1] ,'Price':df.iloc[i,2],'Reviews':df.iloc[i,3] ,'Satisfaction':df.iloc[i,4]} 
        Data2 = Data2.append(temp, ignore_index=True)
Data2

# Export as csv file
Data2.to_csv('Drug_clean.csv', index = False)
