==> 3.Drug_Database.ipynb <==
<b>

<p>
<center>
<font size="6">
Data Mining Project 3
</font>
</center>
</p>

<p>
<center>
<font size="5">
Drug Database
</font>
</center>
</p>

<p>
<center>
<font size="3">
Data Mining, Columbian College of Arts & Sciences, George Washington University
</font>
</center>
</p>

<p>
<center>
<font size="3">
Author: Voratham Tiabrat
</font>
</center>
</p>

</b>

# Part 3 : Drug Database

import pandas as pd
import numpy as np
import seaborn as sns
from matplotlib import pyplot as plt

Drug = pd.read_csv('Drug_clean.csv')
for i in range (Drug.shape[0]) :
    Drug.loc[i,'Condition'] = Drug.loc[i,'Condition'].capitalize()
Drug.head(10)

## 1. Preparing Data

#Change String to Float
Drug['Reviews'] = Drug['Reviews'].astype(float)
Drug['Effective'] = Drug['Effective'].astype(float)
Drug['EaseOfUse'] = Drug['EaseOfUse'].astype(float)
Drug['Satisfaction'] = Drug['Satisfaction'].astype(float)
Drug['Price'] = Drug['Price'].astype(float)

## 1.1 Form of Drug

#Subsetting each form of drug
Tablet = Drug.loc[(Drug.Form == 'Tablet'),]
Capsule = Drug.loc[(Drug.Form == 'Capsule'),]
Cream = Drug.loc[(Drug.Form == 'Cream'),]
Liquid_Drink = Drug.loc[(Drug.Form == 'Liquid (Drink)'),]
Liquid_Inject = Drug.loc[(Drug.Form == 'Liquid (Inject)'),]
Other = Drug.loc[(Drug.Form == 'Other'),]

Form_sum = pd.DataFrame()
Form = [Tablet,Capsule,Cream,Liquid_Drink,Liquid_Inject,Other]
Form_name = ['Tablet','Capsule','Cream','Liquid_Drink','Liquid_Inject','Other']
i = 0
for form in Form :
    sum_temp = form.sum(axis = 0, skipna = True)
    temp = {'Form': Form_name[i],'Reviews_avg':sum_temp[7]/form.shape[0],'Effective_avg':sum_temp[3]/form.shape[0],'EaseOfUse_avg':sum_temp[2]/form.shape[0],'Satisfaction_avg':sum_temp[8]/form.shape[0],'Price_avg':sum_temp[6]/form.shape[0]}
    Form_sum = Form_sum.append(temp, ignore_index=True)
    i += 1

Form_sum = Form_sum.set_index('Form')
Form_sum

## 1.2 Indication

#Subsetting each indication
On_Label = Drug.loc[(Drug.Indication == 'On Label'),]
Off_Label = Drug.loc[(Drug.Indication == 'Off Label'),]

Indication_sum = pd.DataFrame()
Indication = [On_Label,Off_Label]
Indication_name = ['On Label','Off Label']
i = 0
for indic in Indication :
    sum_temp = indic.sum(axis = 0, skipna = True)
    temp = {'Indication': Indication_name[i],'Reviews_avg':sum_temp[7]/indic.shape[0],'Effective_avg':sum_temp[3]/indic.shape[0],'EaseOfUse_avg':sum_temp[2]/indic.shape[0],'Satisfaction_avg':sum_temp[8]/indic.shape[0],'Price_avg':sum_temp[6]/indic.shape[0]}
    Indication_sum = Indication_sum.append(temp, ignore_index=True)
    i += 1

Indication_sum = Indication_sum.set_index('Indication')
Indication_sum

## 1.3 Type

#Subsetting each indication
RX = Drug.loc[(Drug.Type == 'RX'),]
OTC = Drug.loc[(Drug.Type == 'OTC'),]
RX_OTC = Drug.loc[(Drug.Type == 'RX/OTC'),]

Type_sum = pd.DataFrame()
Type = [RX,OTC,RX_OTC]
Type_name = ['RX','OTC','RX/OTC']
i = 0
for typ in Type :
    sum_temp = typ.sum(axis = 0, skipna = True)
    temp = {'Type': Type_name[i],'Reviews_avg':sum_temp[7]/typ.shape[0],'Effective_avg':sum_temp[3]/typ.shape[0],'EaseOfUse_avg':sum_temp[2]/typ.shape[0],'Satisfaction_avg':sum_temp[8]/typ.shape[0],'Price_avg':sum_temp[6]/typ.shape[0]}
    Type_sum = Type_sum.append(temp, ignore_index=True)
    i += 1

Type_sum = Type_sum.set_index('Type')
Type_sum

## 1.4 Condition

Data_grouped_condition = Drug.groupby(['Condition'])
Data_grouped_condition.describe().head()

# Collect the averge of Reviews, Effective, Ease of Use, Satisfaction , Price
Condition_sum = Data_grouped_condition.mean()
Condition_sum

# Making a dictionay of condition
Condition_grouped = list(Data_grouped_condition)
Condition_dict = { Condition_sum.index[i] : i for i in range(0, len(Condition_sum.index) ) }

def menu():
    print()
    print('1. Top Drug for each condition')
    print('2. Overall Drug of each condition')
    print('3. Best Form of Drug for each condition')
    print('4. Best Form of Drug')
    print('5. On Label vs Off Label Drug')
    print('6. RX vs OTC')
    print('7. Form of Drug')
    print('8. Condition name sheet')
    print('q. Quit the program')
    print()
    print('Please type "8" for Condition input')
    print()

def processRequest(selection):
    if selection == '1':
        TopDrug(selection)
    elif selection == '2':
        Overall(selection)    
    elif selection == '3':
        BestFormCondition(selection) 
    elif selection == '4':
        BestForm(selection)
    elif selection == '5':
        OnvsOff(selection)
    elif selection == '6':
        RXvsOTC(selection)
    elif selection == '7':
        DrugForm(selection)  
    elif selection == '8':
        ConditionNameSheet(selection) 
    else:
        return 'q' 

def TopDrug(selection):
    print()
    print('Input the Condition')
    print('For Condition name guideline, please selection "8" on the main menu')
    print()
    Condition = input('Enter Condition : ').capitalize().strip()
    print()
    print('Input the Criteria')
    print('Critiria : Effective, Ease of Use, Satisfaction, Price')
    print()    
    Criteria = input('Enter Criteria : ').lower().replace(" ","")
    print('Input the Minimum of Reviews')
    print()    
    number = input('Enter Number : ').replace(" ","")
        
    try :
        Condition_data = Condition_grouped[Condition_dict[Condition]]
        if Criteria == 'effective':
            Top_Drug = Condition_data[1].sort_values(by='Effective', ascending=False)
        if Criteria == 'easeofuse':
            Top_Drug = Condition_data[1].sort_values(by='EaseOfUse', ascending=False)
        if Criteria == 'satisfaction':
            Top_Drug = Condition_data[1].sort_values(by='Satisfaction', ascending=False)
        if Criteria == 'price':
            Top_Drug = Condition_data[1].sort_values(by='Price')
        Top_Drug = Top_Drug.set_index('Condition')
        Top_Drug = Top_Drug[Top_Drug['Reviews'] > int(number)]
        print('Top ' + number + ' Drug for ' + Condition)
        print(Top_Drug.head(5))
            
            
    except :
        print()
        print('Error : Typing Wrong Input')
                       
def Overall(selection):
    print()
    print('Input the Criteria')
    print('Critiria : Effective, Ease of Use, Satisfaction, Price')
    print()    
    Criteria = input('Enter Criteria : ').lower().replace(" ","")
        
    try :
        Top_Drug = Drug.copy()
        if Criteria == 'effective':
            Cri = 'Effective'
        if Criteria == 'easeofuse':
            Cri = 'EaseOfUse'
        if Criteria == 'satisfaction':
            Cri = 'Satisfaction'
        if Criteria == 'price':
            Cri = 'Price'
            # Some Drug are really expensive, so will show just all drugs that less than 500$
            Top_Drug = Top_Drug[Top_Drug['Price'] < 200]
        plt.figure(figsize=(24,12))
        Plot = sns.boxplot(x="Condition", y=Cri, data=Top_Drug)
        Plot.axes.set_title("Overall Drug by "+Cri,fontsize=30)
        Plot.set_xlabel('Condition',fontsize=20)
        Plot.set_ylabel(Cri,fontsize=20)
        Plot.tick_params(labelsize=15)
        Plot.set_xticklabels(Plot.get_xticklabels(),rotation=90)
        plt.show()
        
    except :
        print()
        print('Error : Typing Wrong Input')
    
    
def BestFormCondition(selection):
    print()
    print('Input the Condition')
    print('For Condition name guideline, please selection "8" on the main menu')
    print()
    Condition = input('Enter Condition : ').capitalize().strip()
    print()
    print('Input the Criteria')
    print('Critiria : Effective, Ease of Use, Satisfaction, Price')
    print()    
    Criteria = input('Enter Criteria : ').lower().replace(" ","")
        
    try :
        Condition_data = Condition_grouped[Condition_dict[Condition]][1]
        if Criteria == 'effective':
            Cri = 'Effective'
        if Criteria == 'easeofuse':
            Cri = 'EaseOfUse'
        if Criteria == 'satisfaction':
            Cri = 'Satisfaction'
        if Criteria == 'price':
            Cri = 'Price'
            # Some Drug are really expensive, so will show just all drugs that less than 500$
            Condition_data = Condition_data[Condition_data['Price'] < 200]
        plt.figure(figsize=(24,12))
        Plot = sns.boxplot(x="Form", y=Cri, data=Condition_data)
        Plot.axes.set_title(Cri + ' of ' + Condition+ "'s Drug",fontsize=30)
        Plot.set_xlabel('Form',fontsize=20)
        Plot.set_ylabel(Cri,fontsize=20)
        Plot.tick_params(labelsize=15)
        Plot.set_xticklabels(Plot.get_xticklabels(),rotation=90)
        plt.show()
            
    except :
        print()
        print('Error : Typing Wrong Input')
        
def BestForm(selection):
    print()
    print('Input the Criteria')
    print('Critiria : Effective, Ease of Use, Satisfaction, Price')
    print()    
    Criteria = input('Enter Criteria : ').lower().replace(" ","")
        
    try :
        Top_Drug = Drug.set_index('Condition')
        if Criteria == 'effective':
            Cri = 'Effective'
        if Criteria == 'easeofuse':
            Cri = 'EaseOfUse'
        if Criteria == 'satisfaction':
            Cri = 'Satisfaction'
        if Criteria == 'price':
            Cri = 'Price'
            # Some Drug are really expensive, so will show just all drugs that less than 500$
            Top_Drug = Top_Drug[Top_Drug['Price'] < 200]
        plt.figure(figsize=(24,12))
        Plot = sns.boxplot(x="Form", y=Cri, data=Top_Drug)
        Plot.axes.set_title("Best Form by "+Cri,fontsize=30)
        Plot.set_xlabel('Form',fontsize=20)
        Plot.set_ylabel(Cri,fontsize=20)
        Plot.tick_params(labelsize=15)
        Plot.set_xticklabels(Plot.get_xticklabels(),rotation=90)
        plt.show()
        print(Form_sum)

    except :
        print()
        print('Error : Typing Wrong Input')

def OnvsOff(selection) :
    print()
    print('On Label medications are FDA approved for the treatment of this condition')
    print('Off Label medications are occasionally used but are not FDA approved for the treatment of this condition')
    print()
    print('Input the Criteria')
    print('Critiria : Effective, Ease of Use, Satisfaction, Price')
    print()    
    Criteria = input('Enter Criteria : ').lower().replace(" ","")
        
    try :
        On_Drug = Drug[Drug['Indication'] == 'On Label']
        Off_Drug = Drug[Drug['Indication'] == 'Off Label']
        Top_Drug = pd.concat([On_Drug,Off_Drug], axis=0)
        if Criteria == 'effective':
            Cri = 'Effective'
        if Criteria == 'easeofuse':
            Cri = 'EaseOfUse'
        if Criteria == 'satisfaction':
            Cri = 'Satisfaction'
        if Criteria == 'price':
            Cri = 'Price'
            # Some Drug are really expensive, so will show just all drugs that less than 500$
            Top_Drug = Top_Drug[Top_Drug['Price'] < 200]
        plt.figure(figsize=(24,12))
        Plot = sns.boxplot(x="Indication", y=Cri, data=Top_Drug)
        Plot.axes.set_title("Indication by "+Cri,fontsize=30)
        Plot.set_xlabel('Indication',fontsize=20)
        Plot.set_ylabel(Cri,fontsize=20)
        Plot.tick_params(labelsize=15)
        Plot.set_xticklabels(Plot.get_xticklabels(),rotation=90)
        plt.show()
        print(Indication_sum)
    
    except :
        print()
        print('Error : Typing Wrong Input')
        
def RXvsOTC(selection) :
    print()
    print('Rx medications require a prescription')
    print('OTC medications may be purchased Over the Counter')
    print()
    print('Input the Criteria')
    print('Critiria : Effective, Ease of Use, Satisfaction, Price')
    print()    
    Criteria = input('Enter Criteria : ').lower().replace(" ","")
        
    try :
        RX_Drug = Drug[Drug['Type'] == 'RX']
        OTC_Drug = Drug[Drug['Type'] == 'OTC']
        RXOTC_Drug = Drug[Drug['Type'] == 'RX/OTC']
        Top_Drug = pd.concat([RX_Drug,OTC_Drug,RXOTC_Drug], axis=0)
        if Criteria == 'effective':
            Cri = 'Effective'
        if Criteria == 'easeofuse':
            Cri = 'EaseOfUse'
        if Criteria == 'satisfaction':
            Cri = 'Satisfaction'
        if Criteria == 'price':
            Cri = 'Price'
            # Some Drug are really expensive, so will show just all drugs that less than 500$
            Top_Drug = Top_Drug[Top_Drug['Price'] < 200]
        plt.figure(figsize=(24,12))
        Plot = sns.boxplot(x="Type", y=Cri, data=Top_Drug)
        Plot.axes.set_title("Type by "+Cri,fontsize=30)
        Plot.set_xlabel('Type',fontsize=20)
        Plot.set_ylabel(Cri,fontsize=20)
        Plot.tick_params(labelsize=15)
        Plot.set_xticklabels(Plot.get_xticklabels(),rotation=90)
        plt.show()
        print(Type_sum)
    
    except :
        print()
        print('Error : Typing Wrong Input')
        
def DrugForm(selection) :
    explode = (0.05,0.05,0.05,0.05,0.05,0.05)
    colors = ['#ff9999','#66b3ff','#99ff99','#ffcc99','#fed0fc','#ceaefa']

    Data_group = Drug.groupby(['Form'])
    Data_size = Data_group.size()
    Data_size = pd.DataFrame(Data_size)
    Data_size.columns = ['Total']
    DrugOrder = Data_size.sort_values(by='Total', ascending = False)
    plt.pie(DrugOrder, labels=DrugOrder.index, explode = explode, autopct='%1.1f%%', colors = colors)
    plt.title('Form of Drug')
    plt.legend(loc='best', bbox_to_anchor=(1, 0.5))
    plt.show()
        
def ConditionNameSheet(selection) :
    conditionnamesheet = pd.DataFrame(Condition_sum.index)
    print(conditionnamesheet)

def main():
    print('Welcome to Common Drugs Database')     
    selection = ''
    while selection != 'q':
        menu()
        selection = input('Make a selection: ')
        response = processRequest(selection)
        if response == 'q':
            break
                                 
main()
